Initial work on making timers work on Windows

+++ dependencies/Panacea/Platform/winheaders.h
@@ -18,7 +18,7 @@ typedef struct tagRECT
 #ifndef GUID_DEFINED
 #define GUID_DEFINED 1
 typedef struct _GUID {
-	unsigned long  Data1;
+	unsigned int   Data1;
 	unsigned short Data2;
 	unsigned short Data3;
 	unsigned char  Data4[ 8 ];
